requirements.txt
asgiref==3.5.2
astropy==5.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
cftime==1.6.1
cycler==0.11.0
decode==0.7.0
Django==3.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
fonttools==4.34.4
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
morecopy==0.2.4
netCDF4==1.6.0
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.8.0
pyerfa==2.0.0.1
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing-extensions==3.10.0.2
xarray==0.20.2
xarray-dataclasses==1.2.0
